Added contains() function
This function will search through the list and see if we have that
data in the list. Inorder to acomplish this we need the client to
implement a compare_to() function since we have no idea what these
pointers point to. The compare_to() function is alot like Java's.
Will be adding an example in the main file.