Added the tree data structure
It is actually a binary search tree. This is the most used tree out of
all the types of trees so I decided to implement it. It still has alot
of work to be done before it can be considered complete (like adding
a delete function) but you have to start somewhere!